frontend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "oauthlib>=3.2.2",
    "pyjwt>=2.10.1",
    "ffmpeg-python>=0.2.0",
    "requests>=2.32.4",
    "ffmpeg>=1.4",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "flask-migrate>=4.1.0",
    "python-dotenv>=1.1.0",
    "boto3>=1.38.36",
]
